debugged backend and frontend
The code works now, but the response is always negative. I believe this is a problem with the model. Although I didn't check the models, I noticed strange behavior. I checked here to make sure the input properly changes the test data. I can always recheck, but we need to really need to train the model properly.